Split, refactored and extended view projection samples
